pyproject.toml
[project]
name = "flood-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "albumentations>=2.0.8",
    "hf-xet>=1.1.10",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "segmentation-models-pytorch>=0.5.0",
]
